requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.7
beartype==0.22.9
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.4
cyclopts==4.5.1
deprecation==2.1.0
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastmcp==2.14.5
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai>=0.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lupa==2.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
multidict==6.7.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
packaging==26.0
pathable==0.4.4
pathvalidate==3.3.1
platformdirs==4.5.1
postgrest==2.27.2
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.5
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.17.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pyperclip==1.11.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
realtime==2.27.2
redis==7.1.1
referencing==0.36.2
requests==2.32.5
rich==14.3.2
rich-rst==1.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tqdm==4.67.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
weaviate-client>=4.19.2
validators>=0.35.0